refactor: rename files in `transactions` directory
